ref: added resource kind to PieceProcessor isSupported method
