+++ api/apps/geoprocessing/src/export/pieces-exporters/scenario-export-config.piece-exporter.ts
@@ -2,7 +2,7 @@ import { geoprocessingConnections } from '@marxan-geoprocessing/ormconfig';
 import { ClonePiece, ExportJobInput, ExportJobOutput } from '@marxan/cloning';
 import { ResourceKind } from '@marxan/cloning/domain';
 import { ClonePieceRelativePaths } from '@marxan/cloning/infrastructure/clone-piece-data';
-import { ProjectMetadataContent } from '@marxan/cloning/infrastructure/clone-piece-data/project-metadata';
+import { ScenarioExportConfigContent } from '@marxan/cloning/infrastructure/clone-piece-data/export-config';
 import { FileRepository } from '@marxan/files-repository';
 import { Injectable } from '@nestjs/common';
 import { InjectEntityManager } from '@nestjs/typeorm';
@@ -16,7 +16,7 @@ import {
 
 @Injectable()
 @PieceExportProvider()
-export class ProjectMetadataPieceExporter implements ExportPieceProcessor {
+export class ScenarioExportConfigPieceExporter implements ExportPieceProcessor {
   constructor(
     private readonly fileRepository: FileRepository,
     @InjectEntityManager(geoprocessingConnections.apiDB)
@@ -24,29 +24,33 @@ export class ProjectMetadataPieceExporter implements ExportPieceProcessor {
   ) {}
 
   isSupported(piece: ClonePiece, kind: ResourceKind): boolean {
-    return (
-      piece === ClonePiece.ProjectMetadata && kind === ResourceKind.Project
-    );
+    return piece === ClonePiece.ExportConfig && kind === ResourceKind.Scenario;
   }
 
   async run(input: ExportJobInput): Promise<ExportJobOutput> {
-    const [projectData]: {
+    const [scenario]: {
       name: string;
+      project_id: string;
       description: string;
     }[] = await this.entityManager.query(
-      `SELECT projects.name, projects.description FROM projects WHERE projects.id = $1`,
+      `
+       SELECT name, project_id, description FROM scenarios where id = $1
+    `,
       [input.resourceId],
     );
 
-    if (!projectData) {
-      throw new Error(
-        `${ProjectMetadataPieceExporter.name} - Project ${input.resourceId} does not exist.`,
-      );
+    if (!scenario) {
+      throw new Error(`Scenario with ID ${input.resourceId} not found`);
     }
 
-    const fileContent: ProjectMetadataContent = {
-      name: projectData.name,
-      description: projectData.description,
+    const fileContent: ScenarioExportConfigContent = {
+      version: `0.1.0`,
+      name: scenario.name,
+      description: scenario.description,
+      projectId: scenario.project_id,
+      resourceKind: input.resourceKind,
+      resourceId: input.resourceId,
+      pieces: input.allPieces,
     };
 
     const outputFile = await this.fileRepository.save(
@@ -56,7 +60,7 @@ export class ProjectMetadataPieceExporter implements ExportPieceProcessor {
 
     if (isLeft(outputFile)) {
       throw new Error(
-        `${ProjectMetadataPieceExporter.name} - Project - couldn't save file - ${outputFile.left.description}`,
+        `${ScenarioExportConfigPieceExporter.name} - Scenario - couldn't save file - ${outputFile.left.description}`,
       );
     }
 
@@ -65,8 +69,7 @@ export class ProjectMetadataPieceExporter implements ExportPieceProcessor {
       uris: [
         {
           uri: outputFile.right,
-          relativePath:
-            ClonePieceRelativePaths[ClonePiece.ProjectMetadata].projectMetadata,
+          relativePath: ClonePieceRelativePaths[ClonePiece.ExportConfig].config,
         },
       ],
     };
