+++ api/libs/cloning/src/piece-processor.port.ts
@@ -1,7 +1,8 @@
+import { ResourceKind } from './domain';
 import { ClonePiece } from './domain/clone-piece';
 
 export abstract class PieceProcessor<I, O> {
   abstract run(input: I): Promise<O>;
 
-  abstract isSupported(piece: ClonePiece): boolean;
+  abstract isSupported(piece: ClonePiece, kind: ResourceKind): boolean;
 }
